+++ example/lib/pages/home.dart
@@ -1,4 +1,5 @@
-import 'package:example/generated/l10n.dart';
+import 'package:example/l10n/generated/app_l10n.dart';
+import 'package:jaspr/dom.dart';
 import 'package:jaspr/jaspr.dart';
 import 'package:jaspr_localizations/jaspr_localizations.dart';
 
@@ -9,12 +10,22 @@ import '../components/language.dart';
 // - this file and any imported file must be compilable for both server and client environments.
 // - this component and any child components will be built once on the server during pre-rendering and then
 //   again on the client during normal rendering.
-@client
 class Home extends StatefulComponent {
   const Home({super.key});
 
   @override
   State<Home> createState() => HomeState();
+
+  @css
+  static List<StyleRule> get styles => [
+    css('.row').styles(
+      display: Display.flex,
+      padding: Spacing.all(16.px),
+      flexDirection: FlexDirection.row,
+      flexWrap: FlexWrap.wrap,
+      gap: Gap.all(50.px),
+    ),
+  ];
 }
 
 class HomeState extends State<Home> {
@@ -34,19 +45,29 @@ class HomeState extends State<Home> {
 
   @override
   Component build(BuildContext context) {
-    return section([
-      img(src: 'images/logo.svg', width: 80),
-      h1([text('Welcome')]),
-      p([text('You successfully create a new Jaspr site with jaspr_localizations.')]),
-      div(styles: Styles(height: 100.px), []),
-      const LanguageComponent(),
+    final l10n = AppL10n.of(context)!;
 
-      div(styles: Styles(height: 50.px), []),
-      JasprLocalizations.withLocale(
-        locale: L10nDelegate.supportedLocales.last,
-        child: LanguageComponent(),
-      ),
-      div(styles: Styles(height: 50.px), []),
-    ]);
+    return section(
+      [
+        img(src: 'images/logo.svg', width: 80),
+        h1([Component.text(l10n.welcome)]),
+        p([Component.text(l10n.successCreation)]),
+        div(styles: Styles(height: 100.px), []),
+
+        div(classes: 'row', [
+          const LanguageComponent(),
+
+          JasprLocalizations.withLocale(
+            locale: AppL10nDelegate.supportedLocales.last,
+            child: LanguageComponent(),
+          ),
+        ]),
+
+        div(styles: Styles(height: 50.px), []),
+      ],
+    );
   }
+
+  @css
+  static List<StyleRule> get styles => [];
 }