+++ lib/jaspr_localizations.dart
@@ -61,6 +61,7 @@
 /// import 'package:jaspr_localizations/jaspr_localizations.dart';
 /// import 'package:example/generated/l10n.dart';
 ///
+/// @client
 /// class App extends StatelessComponent {
 ///   @override
 ///   Component build(BuildContext context) {
@@ -183,11 +184,8 @@ library;
 // export 'package:intl/intl.dart';
 
 export 'src/jaspr_l10n_core.dart';
-export 'src/jaspr_l10n_types.dart';
-export 'src/utils/localizations_utils.dart';
 export 'src/utils/current_language.dart';
 export 'src/widgets/jaspr_localization.dart';
 export 'src/base/locale.dart';
 export 'src/base/locale_change_notifier.dart';
 export 'src/base/localizations_delegate.dart';
-export 'src/base/builder.dart';