+++ lib/src/jaspr_l10n_generator.dart
@@ -6,7 +6,6 @@
 // Based on Flutter's gen_l10n.dart but adapted for Jaspr framework
 
 import 'dart:io';
-
 import 'base/file_system.dart';
 import 'base/logger.dart';
 import 'jaspr_l10n_templates.dart';
@@ -77,7 +76,7 @@ class LocalizationsGenerator {
     outputDirectory = fileSystem.directory(outputPathString);
     templateArbFile = fileSystem.file(templateArbPathString);
 
-    final String? templateBundlePathString = templateArbPathString;
+    final String templateBundlePathString = templateArbPathString;
     try {
       templateBundle = AppResourceBundle(templateArbFile);
     } on FileSystemException catch (e) {
@@ -342,6 +341,7 @@ class LocalizationsGenerator {
           if (format != null) {
             return dateFormattingTemplate
                 .replaceAll('@{formatterName}', '${placeholder.name}Formatter')
+                // ignore: unnecessary_brace_in_string_interps
                 .replaceAll('@{constructorCall}', '.${format}')
                 .replaceAll('@{parameters}', 'localeName');
           } else {
@@ -361,6 +361,7 @@ class LocalizationsGenerator {
           final String format = placeholder.format!;
           return numberFormattingTemplate
               .replaceAll('@{formatterName}', '${placeholder.name}Formatter')
+              // ignore: unnecessary_brace_in_string_interps
               .replaceAll('@{constructorCall}', '.${format}')
               .replaceAll('@{parameters}', 'locale: localeName');
         })
@@ -374,9 +375,16 @@ class LocalizationsGenerator {
   String _generateBaseClassMethods() {
     final StringBuffer sb = StringBuffer();
     for (final Message message in allMessages) {
+      String comment =
+          message.description ?? 'No description provided for @{methodName}.';
+      if (message.context != null) {
+        comment = '$comment\n  ///\n  /// context: ${message.context}';
+      }
+
       if (message.templatePlaceholders.isEmpty) {
         sb.writeln(
           baseClassGetterTemplate
+              .replaceAll('@{comment}', comment)
               .replaceAll('@{methodName}', message.resourceId)
               .replaceAll('@{localeVar}', templateBundle.locale.toString())
               .replaceAll('@{message}', message.value),
@@ -390,6 +398,7 @@ class LocalizationsGenerator {
 
         sb.writeln(
           baseClassMethodTemplate
+              .replaceAll('@{comment}', comment)
               .replaceAll('@{methodName}', message.resourceId)
               .replaceAll('@{methodParameters}', methodParameters)
               .replaceAll('@{localeVar}', templateBundle.locale.toString())
@@ -412,7 +421,7 @@ class LocalizationsGenerator {
     final StringBuffer sb = StringBuffer();
     for (final LocaleInfo locale in supportedLocales) {
       final String className =
-          '$classNameString${_camelCase(locale.toString())}';
+          '_$classNameString${_camelCase(locale.toString())}';
       sb.writeln(
         jasprFromSwitchCaseTemplate
             .replaceAll('@{locale}', locale.toString())
@@ -481,7 +490,7 @@ class LocalizationsGenerator {
     // Generate individual locale classes
     for (final LocaleInfo locale in supportedLocales) {
       final String className =
-          '$classNameString${_camelCase(locale.toString())}';
+          '_$classNameString${_camelCase(locale.toString())}';
       final String classContent = jasprFileTemplate
           .replaceAll('@{locale}', locale.toString())
           .replaceAll('@{class}', className)